feature/ created template pick options in account menu

+++ accounts/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.core.exceptions import ValidationError
+from website.models import TemplateChoice, Templates
 from re import match
 
 # Create your views here.
@@ -19,7 +20,7 @@ def username_validator(variable):
     return False
 
 def name_validator(variable):
-    pat = "/[a-z]/gi"
+    pat = "^[a-zA-Z]+$"
     if match(pat, variable):
         return True
     return False
@@ -31,12 +32,28 @@ def accounts(request):
     current_user = request.user.id
     fetch_user = User.objects.filter(id=current_user)
     context = {"account_info": fetch_user}
+    user_template = Templates.objects.get(user_id=current_user)
     if request.method == "POST":
-        id_to_delete = request.POST["delete-account"]
-        active_user = User.objects.get(id=id_to_delete)
-        # print(f"deleting", active_user)
-        active_user.delete()
-        return redirect("log_in")
+        if "delete-account" in request.POST:
+            id_to_delete = request.POST["delete-account"]
+            active_user = User.objects.get(id=id_to_delete)
+            active_user.delete()
+            return redirect("log_in")
+        elif "orange" in request.POST:
+            template_choice = TemplateChoice.objects.get(id=1)
+            user_template.template_id = template_choice
+            user_template.save()
+            print(user_template.template_id)
+        elif "green" in request.POST:
+            template_choice = TemplateChoice.objects.get(id=2)
+            user_template.template_id = template_choice
+            user_template.save()
+            print(user_template.template_id)
+        elif "purple" in request.POST:
+            template_choice = TemplateChoice.objects.get(id=3)
+            user_template.template_id = template_choice
+            user_template.save()
+            print(user_template.template_id)
     return render(request, "accounts/accounts.html", context)
 
 def log_out(request):
@@ -96,6 +113,8 @@ def sign_up(request):
                             first_name=name_cleaner(first_name),
                             last_name=name_cleaner(last_name))
                         new_user.save()
+                        user_template = Templates.objects.create(user_id=new_user)
+                        user_template.save()
                         return redirect("reg_success")
                     else:
                         name_error = {"name_error": "First and Last name can only contain letters!"}
